pyproject.toml
[project]
name = "test-git"
version = "0.1.7"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
]

[dependency-groups]
dev = [
    "bump-my-version>=1.2.1",
]

github-actions = [
    "pydantic-settings >=2.1.0,<3.0.0",
    "pygithub >=2.3.0,<3.0.0",
    "pyyaml >=5.3.1,<7.0.0",
]
